subscriber/requirements.txt
paho-mqtt
psycopg[binary]
psycopg[pool]
python-dotenv
numpy
geopy
shapely
